moving exception hook to drum.py
